+++ lib/absinthe/phase/subscription/subscribe_self.ex
@@ -6,7 +6,7 @@ defmodule Absinthe.Phase.Subscription.SubscribeSelf do
 
   alias Absinthe.Blueprint
 
-  @spec run(any, Keyword.t()) :: {:ok, Blueprint.t()}
+  @spec run(any, Keyword.t()) :: Phase.result_t()
   def run(blueprint, options) do
     with %{type: :subscription} = op <- Blueprint.current_operation(blueprint) do
       do_subscription(op, blueprint, options)